requirements.txt
# Core Django
django==4.2.17
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.11
dj-database-url==2.3.0


# Server
gunicorn==25.1.0
whitenoise==6.11.0

# Environment
python-dotenv==1.2.1

# Auth / Security
passlib==1.7.4
bcrypt==3.2.2

# AI / ML
tensorflow==2.16.1
transformers==4.44.2
torch==2.4.1
tensorflow-cpu==2.16.1
pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.9.0.80

# Utilities
requests==2.32.3
packaging==26.0
sendgrid==6.11.0
